# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DetectionOutput": {
        "description": "Output for detection.\n\n``masks`` is populated when the model performs instance segmentation\n(boxes + masks).  For detection-only models it is ``None``.\n\nAttributes:\n    boxes: List of detected boxes.\n    scores: List of confidence scores.\n    classes: List of class names associated with each box.\n    masks: Optional instance masks in compressed-RLE format.",
        "properties": {
          "boxes": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Boxes",
            "type": "array"
          },
          "classes": {
            "items": {
              "type": "string"
            },
            "title": "Classes",
            "type": "array"
          },
          "masks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CompressedRLE-Output"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Masks"
          },
          "scores": {
            "items": {
              "type": "number"
            },
            "title": "Scores",
            "type": "array"
          }
        },
        "required": [
          "boxes",
          "scores",
          "classes"
        ],
        "title": "DetectionOutput",
        "type": "object"
      },
      "DetectionRequest": {
        "description": "Request for detection inference.",
        "properties": {
          "boxThreshold": {
            "default": 0.5,
            "title": "Boxthreshold",
            "type": "number"
          },
          "classes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classes"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "format": "path",
                "type": "string"
              }
            ],
            "title": "Image"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "textThreshold": {
            "default": 0.5,
            "title": "Textthreshold",
            "type": "number"
          }
        },
        "required": [
          "image",
          "model"
        ],
        "title": "DetectionRequest",
        "type": "object"
      },
      "DetectionResponse": {
        "description": "Response for detection inference.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DetectionOutput"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "processingTime": {
            "default": 0.0,
            "title": "Processingtime",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "timestamp",
          "metadata",
          "data"
        ],
        "title": "DetectionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NERRequest": {
        "description": "Request for NER inference.",
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "model"
        ],
        "title": "NERRequest",
        "type": "object"
      },
      "NERResponse": {
        "description": "Response for NER inference.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NEROutput"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "processingTime": {
            "default": 0.0,
            "title": "Processingtime",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "timestamp",
          "metadata",
          "data"
        ],
        "title": "NERResponse",
        "type": "object"
      },
...
      "SegmentationRequest": {
        "description": "Request for segmentation inference.",
        "properties": {
          "boxes": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boxes"
          },
          "highResolutionFeatures": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NDArrayFloat"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highresolutionfeatures"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "format": "path",
                "type": "string"
              },
              {
                "format": "binary",
                "type": "string"
              }
            ],
            "title": "Image"
          },
          "imageEmbedding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NDArrayFloat"
              },
              {
                "type": "null"
              }
            ]
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "maskInput": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NDArrayFloat"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "multimaskOutput": {
            "default": true,
            "title": "Multimaskoutput",
            "type": "boolean"
          },
          "numMultimaskOutputs": {
            "default": 3,
            "title": "Nummultimaskoutputs",
            "type": "integer"
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "items": {
                      "type": "integer"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "resetPredictor": {
            "default": true,
            "title": "Resetpredictor",
            "type": "boolean"
          },
          "returnImageEmbedding": {
            "default": false,
            "title": "Returnimageembedding",
            "type": "boolean"
          },
          "returnLogits": {
            "default": false,
            "title": "Returnlogits",
            "type": "boolean"
          }
        },
        "required": [
          "image",
          "model"
        ],
        "title": "SegmentationRequest",
        "type": "object"
      },
      "SegmentationResponse": {
        "description": "Response for segmentation inference.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SegmentationOutput"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "processingTime": {
            "default": 0.0,
            "title": "Processingtime",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "timestamp",
          "metadata",
          "data"
        ],
        "title": "SegmentationResponse",
        "type": "object"
      },
...
      "TrackingRequestV1": {
        "description": "Video-tracking request with nested keyframe prompts.\n\nMedia is passed by value (URL/base64/frame list/path within an allowed media root);\ndataset-reference resolution belongs in the caller (the Pixano backend).",
        "properties": {
          "frameIndexes": {
            "items": {
              "type": "integer"
            },
            "title": "Frameindexes",
            "type": "array"
          },
          "interval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrackingInterval"
              },
              {
                "type": "null"
              }
            ]
          },
          "keyframes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TrackingKeyframeV1"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyframes"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "objectsIds": {
            "items": {
              "type": "integer"
            },
            "title": "Objectsids",
            "type": "array"
          },
          "propagate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Propagate"
          },
          "video": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "format": "path",
                      "type": "string"
                    },
                    {
                      "format": "binary",
                      "type": "string"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "format": "path",
                "type": "string"
              },
              {
                "format": "binary",
                "type": "string"
              }
            ],
            "title": "Video"
          }
        },
        "required": [
          "model",
          "video",
          "objectsIds",
          "frameIndexes"
        ],
        "title": "TrackingRequestV1",
        "type": "object"
      },
      "TrackingResponse": {
        "description": "Response for tracking inference.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TrackingOutput"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "processingTime": {
            "default": 0.0,
            "title": "Processingtime",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "timestamp",
          "metadata",
          "data"
        ],
        "title": "TrackingResponse",
        "type": "object"
      },
...
      "VLMRequest": {
        "description": "Request for VLM inference.",
        "properties": {
          "images": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "format": "path",
                      "type": "string"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "maxNewTokens": {
            "title": "Maxnewtokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              }
            ],
            "title": "Prompt"
          },
          "temperature": {
            "default": 1.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "prompt",
          "maxNewTokens",
          "model"
        ],
        "title": "VLMRequest",
        "type": "object"
      },
      "VLMResponse": {
        "description": "Response for VLM inference.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VLMOutput"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "processingTime": {
            "default": 0.0,
            "title": "Processingtime",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "timestamp",
          "metadata",
          "data"
        ],
        "title": "VLMResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Pixano Inference API powered by Ray Serve",
    "title": "Pixano Inference (Ray)",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/inference/detection": {
      "post": {
        "operationId": "detection_v1_inference_detection_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detection",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/detection/binary": {
      "post": {
        "operationId": "detection_binary_v1_inference_detection_binary_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detection Binary",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/ner": {
      "post": {
        "operationId": "ner_v1_inference_ner_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NERRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NERResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ner",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/segmentation": {
      "post": {
        "operationId": "segmentation_v1_inference_segmentation_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Segmentation",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/segmentation/binary": {
      "post": {
        "operationId": "segmentation_binary_v1_inference_segmentation_binary_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Segmentation Binary",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/tracking": {
      "post": {
        "operationId": "tracking_v1_inference_tracking_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tracking",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/tracking/binary": {
      "post": {
        "operationId": "tracking_binary_v1_inference_tracking_binary_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tracking Binary",
        "tags": [
          "inference"
        ]
      }
    },
...
    "/v1/inference/vlm": {
      "post": {
        "operationId": "vlm_v1_inference_vlm_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VLMRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VLMResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vlm",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/inference/vlm/binary": {
      "post": {
        "operationId": "vlm_binary_v1_inference_vlm_binary_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VLMResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vlm Binary",
        "tags": [
          "inference"
        ]
      }
    },
...
  }
}